+++ src/Model/Request.php
@@ -120,7 +120,7 @@ class Request
     private ?string $details = null;
 
     /**
-     * @ORM\Column(type="text", length=65535, nullable=true)
+     * @ORM\Column(type="text", length=16777215, nullable=true)
      */
     private ?string $killMail = null;